Move limb into draw
Combines limb into the draw module (as well as the tests) and removes the old limb and test_limb modules
