+++ sunpy/visualization/__init__.py
@@ -1,2 +1,2 @@
-from sunpy.visualization.limb import *
+from sunpy.visualization.draw import *
 from sunpy.visualization.visualization import *
